Refactor CLI structure by removing outdated routes and adding new ingest functionality
- Deleted `es_templates.route.ts` and `observe_templates.route.ts` as they are no longer needed.
- Introduced new ingest functionality in `ingest.ts` and various `ingest-ska` modules to handle data ingestion from the SKA API.
- Added a README file for the CLI to provide usage instructions and available datasets.
- Updated `ingest-wab` functionality to improve data processing and indexing.
