+++ apps/api/src/cli/ingest-ska/get-ska-documents.ts
@@ -1,9 +1,9 @@
 import { DATA_SOURCES } from '@shared/constants';
-import { isObjectEmpty } from '@lib/isObjectEmpty';
+import { isObjectEmpty } from '@shared/utils/isObjectEmpty';
 import * as jsonld from 'jsonld';
 import contexts from 'jsonld-contexts';
 
-const SKA_API = DATA_SOURCES.filter((service) => service.name === 'spes')[0].url;
+const SKA_API = DATA_SOURCES.filter((service) => service.name === 'ska')[0].url;
 
 export async function getSkaDocuments() {
   const query = `
